backend/requirements/base.txt
# --- Core framework ---
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
django-filter==24.3
django-environ==0.11.2
drf-spectacular==0.27.2

# --- Database ---
psycopg[binary]==3.2.3

# --- Celery ---
celery==5.4.0
redis==5.2.0
django-celery-beat==2.7.0
django-celery-results==2.5.1

# --- Document generation ---
python-docx==1.1.2
reportlab==4.2.5
qrcode==8.0
num2words==0.5.14
openpyxl==3.1.5
Pillow==11.0.0

# --- HTTP client (for Eskiz SMS) ---
httpx==0.28.1

# --- Server ---
gunicorn==23.0.0
